requirements.txt
requests==2.5.3
six==1.9.0
tapioca-wrapper==0.2.1
